whisper-service/requirements.txt
flask
openai-whisper
ffmpeg-python
flask-cors
deep-translator